o Resolved MNG-692, although I think a more general solution is required;
o Fixed another boolean-logic error in the ear plugin.
  Stephane: (x != a) || (x != b) is always true, unless a == b.. ;-)


git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk@227198 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-plugins/maven-ear-plugin/src/main/java/org/apache/maven/plugin/ear/EarMojo.java
@@ -119,6 +119,15 @@ public class EarMojo
                 EarModule module = (EarModule) iter.next();
                 getLog().info( "Copying artifact[" + module + "] to[" + module.getUri() + "]" );
                 File destinationFile = buildDestinationFile( getBuildDir(), module.getUri() );
+
+                File sourceFile = module.getArtifact().getFile();
+
+                if ( !sourceFile.isFile() )
+                {
+                    throw new MojoExecutionException( "Cannot copy a directory: " + sourceFile.getAbsolutePath() + 
+                        "; Did you package/install " + module.getArtifact().getId() + "?" );
+                }
+
                 FileUtils.copyFile( module.getArtifact().getFile(), destinationFile );
             }
         }
@@ -158,6 +167,8 @@ public class EarMojo
 
             archiver.getArchiver().addDirectory( getBuildDir() );
             archiver.createArchive( getProject(), archive );
+
+            project.getArtifact().setFile( earFile );
         }
         catch ( Exception e )
         {
